fix(logging): Rename logging to log, avoids circular import
The existance of the spex.logging and the inhereted python logging
module clashes and can result in circular imports, especially when
running debugpy.
