+++ src/spex/jsonspec/extractors/structtable.py
@@ -17,7 +17,7 @@ from spex.jsonspec.defs import (
 from spex.jsonspec.extractors.figure import FigureExtractor, RowErrPolicy
 from spex.jsonspec.extractors.helpers import content_extract_brief, validate_label
 from spex.jsonspec.lint import LintErr
-from spex.logging import logger
+from spex.log import logger
 from spex.xml import Element, XmlUtils, Xpath
