+++ src/spex/log/__init__.py
@@ -2,8 +2,10 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-from loguru import logger
+from loguru import logger as _logger
 
-from spex.logging._log_config import ULog
+from spex.log._log_config import ULog
+
+logger = _logger
 
 __all__ = ["logger", "ULog"]
